requirements.txt
PyPDF2
python-docx
spacy
pytest
fastapi
uvicorn[standard]
pydantic
